Addition of subjective pronouns.
As per my issue post https://github.com/Alheimsins/b5-johnson-120-ipip-neo-pi-r/issues/204 questions assume the presence of a prononoun preceding the question. I noticed this pronoun is excluded in the downstream application as well, so here I attempted to find them for each language. There are probably a lot of mistakes and these translations need to be looked at in the context of the questions.

Worth noting that the translations are very mixed between including a pronoun and excluding it. For example: Swedish, Norwegian, Danish, English, Dutch, Italian and Portuguese all seem to exclude them, while Icelandic, German, Estonian and French all seem to include them. Ideally this is standardised in future translations and edits.

This change makes them accessible under the getInfo() call.